# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Documentação da API Autoconf Vehicles'
  description: 'API para um SaaS multiusuário de gestão de veículos: autenticação, cadastro de veículos e gestão de imagens com capa única.'
  version: 1.0.0
servers:
  - url: 'http://localhost:8000'
...
paths:
  /api/auth/register:
    post:
      summary: 'Create a new user account.'
      operationId: createANewUserAccount
      description: "This does not also start an authenticated session: registration and\nlogin are kept as two separate, explicit steps, matching the plain\n`POST /auth/register` contract this endpoint promises. `User::create()`\nonly ever receives the request's own validated fields, so a payload\nthat also sends `is_admin` (or any other attribute) has no way to\ninfluence the created record."
      parameters: []
      responses:
        201:
          description: 'Conta criada com sucesso. Note que isso não autentica o usuário — é necessário chamar `POST /auth/login` em seguida.'
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 1
                  name: 'Maria Souza'
                  email: maria.souza@example.com
                  created_at: '2026-08-06T12:00:00.000000Z'
                  updated_at: '2026-08-06T12:00:00.000000Z'
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: 'Maria Souza'
                  email:
                    type: string
                    example: maria.souza@example.com
                  created_at:
                    type: string
                    example: '2026-08-06T12:00:00.000000Z'
                  updated_at:
                    type: string
                    example: '2026-08-06T12:00:00.000000Z'
        422:
          description: 'Payload inválido — e-mail já cadastrado, senha e confirmação não coincidem, ou campo obrigatório ausente.'
          content:
            application/json:
              schema:
                type: object
                example:
                  type: 'about:blank'
                  title: 'Unprocessable Content'
                  status: 422
                  detail: 'Já existe uma conta cadastrada com esse e-mail. (and 1 more error)'
                  instance: /api/auth/register
                  errors:
                    email:
                      - 'Já existe uma conta cadastrada com esse e-mail.'
                    password:
                      - 'A confirmação de senha não corresponde.'
                properties:
                  type:
                    type: string
                    example: 'about:blank'
                  title:
                    type: string
                    example: 'Unprocessable Content'
                  status:
                    type: integer
                    example: 422
                  detail:
                    type: string
                    example: 'Já existe uma conta cadastrada com esse e-mail. (and 1 more error)'
                  instance:
                    type: string
                    example: /api/auth/register
                  errors:
                    type: object
                    properties:
                      email:
                        type: array
                        example:
                          - 'Já existe uma conta cadastrada com esse e-mail.'
                        items:
                          type: string
                      password:
                        type: array
                        example:
                          - 'A confirmação de senha não corresponde.'
                        items:
                          type: string
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Nome completo do usuário.'
                  example: 'Maria Souza'
                email:
                  type: string
                  description: 'E-mail único do usuário.'
                  example: maria.souza@example.com
                password:
                  type: string
                  description: 'Senha do usuário, sujeita às regras padrão de senha do Laravel.'
                  example: Str0ng!Passw0rd
                password_confirmation:
                  type: string
                  description: 'Deve ser idêntica a `password`.'
                  example: Str0ng!Passw0rd
              required:
                - name
                - email
                - password
                - password_confirmation
      security: []
  /api/auth/login:
    post:
      summary: "Authenticate a user for the SPA's stateful (cookie) session."
      operationId: authenticateAUserForTheSPAsStatefulcookieSession
      description: "`ValidationException::withMessages()` on a bad attempt is deliberate,\nnot just convenient: this app's exception renderer already turns any\n`ValidationException` into the same `application/problem+json` 422\nenvelope used by Form Request validation, so a credentials failure\nlooks identical to any other validation failure to API consumers\nwithout the controller having to build that response by hand."
      parameters: []
      responses:
        200:
          description: 'Login bem-sucedido. A resposta carrega o cookie de sessão `HttpOnly`; o corpo é o usuário autenticado.'
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 1
                  name: 'Maria Souza'
                  email: maria.souza@example.com
                  created_at: '2026-08-06T12:00:00.000000Z'
                  updated_at: '2026-08-06T12:00:00.000000Z'
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: 'Maria Souza'
                  email:
                    type: string
                    example: maria.souza@example.com
                  created_at:
                    type: string
                    example: '2026-08-06T12:00:00.000000Z'
                  updated_at:
                    type: string
                    example: '2026-08-06T12:00:00.000000Z'
        422:
          description: 'Credenciais inválidas. Mesmo formato `application/problem+json` usado por qualquer outra falha de validação (422).'
          content:
            application/json:
              schema:
                type: object
                example:
                  type: 'about:blank'
                  title: 'Unprocessable Content'
                  status: 422
                  detail: 'Essas credenciais não correspondem aos nossos registros.'
                  instance: /api/auth/login
                  errors:
                    email:
                      - 'Essas credenciais não correspondem aos nossos registros.'
                properties:
                  type:
                    type: string
                    example: 'about:blank'
                  title:
                    type: string
                    example: 'Unprocessable Content'
                  status:
                    type: integer
                    example: 422
                  detail:
                    type: string
                    example: 'Essas credenciais não correspondem aos nossos registros.'
                  instance:
                    type: string
                    example: /api/auth/login
                  errors:
                    type: object
                    properties:
                      email:
                        type: array
                        example:
                          - 'Essas credenciais não correspondem aos nossos registros.'
                        items:
                          type: string
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'E-mail cadastrado.'
                  example: maria.souza@example.com
                password:
                  type: string
                  description: 'Senha da conta.'
                  example: Str0ng!Passw0rd
              required:
                - email
                - password
      security: []
  /api/auth/me:
    get:
      summary: 'Return the currently authenticated user.'
      operationId: returnTheCurrentlyAuthenticatedUser
      description: "This route sits behind `auth:sanctum`, so a guest never reaches this\nmethod at all: Sanctum's own guard throws `AuthenticationException`\nfirst, which the app's exception renderer already turns into a\nproblem+json 401 (wired in `bootstrap/app.php`)."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 1
                  name: 'Maria Souza'
                  email: maria.souza@example.com
                  created_at: '2026-08-06T12:00:00.000000Z'
                  updated_at: '2026-08-06T12:00:00.000000Z'
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: 'Maria Souza'
                  email:
                    type: string
                    example: maria.souza@example.com
                  created_at:
                    type: string
                    example: '2026-08-06T12:00:00.000000Z'
                  updated_at:
                    type: string
                    example: '2026-08-06T12:00:00.000000Z'
        401:
          description: 'Nenhuma sessão válida (cookie ausente/expirado, ou `X-XSRF-TOKEN` ausente/incorreto).'
          content:
            application/json:
              schema:
                type: object
                example:
                  type: 'about:blank'
                  title: Unauthorized
                  status: 401
                  detail: Unauthenticated.
                  instance: /api/auth/me
                properties:
                  type:
                    type: string
                    example: 'about:blank'
                  title:
                    type: string
                    example: Unauthorized
                  status:
                    type: integer
                    example: 401
                  detail:
                    type: string
                    example: Unauthenticated.
                  instance:
                    type: string
                    example: /api/auth/me
      tags:
        - Auth
  /api/auth/logout:
    post:
      summary: "End the SPA's authenticated session."
      operationId: endTheSPAsAuthenticatedSession
      description: "The three steps mirror `login()`'s own care around sessions, in\nreverse: log the guard out, destroy the session data, and rotate the\nCSRF token, so neither the session id nor the CSRF token an attacker\nobserved before logout stays valid afterwards. `204 No Content` is\nused rather than `200` because there is no representation to return\nfor a logout."
      parameters: []
      responses:
        204:
          description: 'Sessão, dados de sessão e token CSRF invalidados com sucesso.'
          content:
            text/plain:
              schema:
                type: string
                example: ''
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  type: 'about:blank'
                  title: Unauthorized
                  status: 401
                  detail: Unauthenticated.
                  instance: /api/auth/logout
                properties:
                  type:
                    type: string
                    example: 'about:blank'
                  title:
                    type: string
                    example: Unauthorized
                  status:
                    type: integer
                    example: 401
                  detail:
                    type: string
                    example: Unauthenticated.
                  instance:
                    type: string
                    example: /api/auth/logout
      tags:
        - Auth
...